Use newtype wrapper around Project + Pretty instance
